+++ docs/conf.py
@@ -12,6 +12,7 @@
 # serve to show the default.
 
 import os
+import pkg_resources
 import sys
 
 # abspath because this could be __main__, in which case it may not
@@ -73,9 +74,9 @@ copyright = '2015-2017 — Mozilla Services'
 # built documents.
 #
 # The short X.Y version.
-version = '8.3'
 # The full version, including alpha/beta/rc tags.
-release = '8.3.0'
+version = pkg_resources.get_distribution("kinto").version
+release = ".".join(version.split(".")[:2])
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.