+++ kinto/plugins/accounts/__init__.py
@@ -2,6 +2,7 @@ from kinto.authorization import PERMISSIONS_INHERITANCE_TREE
 from pyramid.exceptions import ConfigurationError
 
 ACCOUNT_CACHE_KEY = 'accounts:{}:verified'
+ACCOUNT_POLICY_NAME = 'account'
 
 
 def includeme(config):