+++ tests/core/resource/test_filter.py
@@ -16,8 +16,9 @@ class FilteringTest(BaseTest):
             {'title': 'MoFo', 'status': 0, 'favorite': False, 'sometimes': None},
             {'title': 'MoFo', 'status': 1, 'favorite': True, 'fib': [1, 2, 3]},
             {'title': 'MoFo', 'status': 2, 'favorite': False, 'fib': [3, 5, 8]},
-            {'title': 'Foo', 'status': 3, 'favorite': False},
-            {'title': 'Bar', 'status': 3, 'favorite': False, 'sometimes': 'present'},
+            {'title': 'Foo', 'status': 3, 'favorite': False, "aliases": [{"ll": "ls -l"}]},
+            {'title': 'Bar', 'status': 3, 'favorite': False, 'sometimes': 'present',
+             "aliases": [{"ll": "ls -l"}, {"rm": "rm -i"}]},
         ]
         for r in records:
             self.model.create_record(r)
@@ -219,33 +220,65 @@ class FilteringTest(BaseTest):
         self.assertRaises(httpexceptions.HTTPBadRequest,
                           self.resource.collection_get)
 
-    def test_contains_with_strings(self):
+    def test_contains_can_filter_with_one_string(self):
+        self.validated['querystring'] = {'contains_colors': ["red"]}
+        result = self.resource.collection_get()
+        values = [item['colors'] for item in result['data']]
+        assert len(values) == 1
+        for value in values:
+            assert "red" in value
+
+    def test_contains_can_filter_with_one_object(self):
+        self.validated['querystring'] = {'contains_aliases': [{"ll": "ls -l"}]}
+        result = self.resource.collection_get()
+        values = [item['aliases'] for item in result['data']]
+        assert len(values) == 2
+        for value in values:
+            assert {"ll": "ls -l"} in value
+
+    def test_contains_can_filter_with_list_of_strings(self):
         self.validated['querystring'] = {'contains_colors': ["red", "blue"]}
         result = self.resource.collection_get()
         values = [item['colors'] for item in result['data']]
+        assert len(values) == 1
         for value in values:
             assert "red" in value and "blue" in value
 
-    def test_contains_with_integer(self):
+    def test_contains_can_filter_with_an_integer(self):
         self.validated['querystring'] = {'contains_fib': [3]}
         result = self.resource.collection_get()
         values = [item['fib'] for item in result['data']]
+        assert len(values) == 2
         for value in values:
             assert 3 in value
 
-    def test_contains_any_with_strings(self):
-        self.validated['querystring'] = {'contains_colors': ["red", "blue"]}
+    def test_contains_any_can_filter_with_a_list_of_strings(self):
+        self.validated['querystring'] = {'contains_any_colors': ["red", "blue"]}
         result = self.resource.collection_get()
         values = [item['colors'] for item in result['data']]
+        assert len(values) == 2
         for value in values:
             assert 'red' in value or 'blue' in value
 
-    def test_contains_any_with_integer(self):
-        self.validated['querystring'] = {'contains_fib': [3, 5]}
+    def test_contains_any_can_filter_with_a_list_of_integers(self):
+        self.validated['querystring'] = {'contains_any_fib': [3, 5]}
         result = self.resource.collection_get()
         values = [item['fib'] for item in result['data']]
+        assert len(values) == 2
         for value in values:
             assert 3 in value or 5 in value
+
+    def test_contains_fails_on_a_non_sequence_record_value(self):
+        self.validated['querystring'] = {'contains_favorite': [True]}
+        result = self.resource.collection_get()
+        values = result['data']
+        assert len(values) == 0
+
+    def test_contains_any_fails_on_a_non_sequence_record_value(self):
+        self.validated['querystring'] = {'contains_any_favorite': [True]}
+        result = self.resource.collection_get()
+        values = result['data']
+        assert len(values) == 0
 
 
 class SubobjectFilteringTest(BaseTest):