+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ This document describes changes between each past release.
 
 - Introduce ``contains`` and ``contains_any`` filter operators (fixes #343).
 
+**Documentation**
+
+- Version number is taken from package in order to ease release process
+
 
 8.3.0 (2018-04-06)
 ------------------