+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ Changelog
 This document describes changes between each past release.
 
 
+9.1.0 (unreleased)
+------------------
+
+- Nothing changed yet.
+
+
 9.0.0 (2018-04-25)
 ------------------